requirements.txt
# Web Scraping Dependencies
beautifulsoup4==4.12.3
requests==2.32.3
scrapy==2.11.2
selenium==4.23.1
lxml==5.2.2

# Data Processing Dependencies
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5

# NLP Processing Dependencies
spacy==3.7.5
openai==1.40.0
python-dotenv==1.0.1

# Web Interface Dependencies
streamlit==1.37.1
plotly==5.23.0
altair==5.3.0

# Database
sqlite3  # Built-in with Python

# Additional Utilities
python-dateutil==2.9.0
validators==0.33.0